identifier/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pillow==10.4.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
faiss-cpu==1.8.0.post1
torch==2.2.2
torchvision==0.17.2
transformers==4.45.2

# Test dependencies
pytest>=8.0.0
httpx>=0.27.0
